refactor: move domain user library
